glt - Update DELETE /api/courses/user endpoint to delete section/lectures together
I updated the endpoint to delete all PSCourse entries in the database
that have enroll codes related to the argument enroll code.

We have a new way in which EntityNotFoundError can arise, but it is
not an intuitive for the caller. (Not relevant as long as database is
in a consistent state.)

When the UCSB API delivers bad data, it defaults to the previous
behavior.  This and any other processing errors cause the previous
behavior.